refactor: replace NoInternetScreen with ErrorView and update error handling in TransactionDetails
