+++ wallet_presentation/src/commonMain/kotlin/net/thechance/mena/wallet/presentation/screen/transaction_details/TransactionDetailsScreen.kt
@@ -18,7 +18,7 @@ import net.thechance.mena.designsystem.presentation.component.appBar.AppBar
 import net.thechance.mena.designsystem.presentation.component.icon.Icon
 import net.thechance.mena.designsystem.presentation.theme.theme.MenaTheme
 import net.thechance.mena.designsystem.presentation.theme.theme.Theme
-import net.thechance.mena.wallet.presentation.component.NoInternetScreen
+import net.thechance.mena.wallet.presentation.component.ErrorView
 import net.thechance.mena.wallet.presentation.component.SnackBarContainer
 import net.thechance.mena.wallet.presentation.component.WalletScaffold
 import net.thechance.mena.wallet.presentation.screen.transaction_details.TransactionDetailsScreenState.TransactionDetailsUiState
@@ -55,10 +55,7 @@ fun TransactionDetailsScreen(
         }
     )
 
-    TransactionDetailsScreenContent(
-        state = state,
-        interactionListener = viewModel
-    )
+    TransactionDetailsScreenContent(state = state, interactionListener = viewModel)
 }
 
 @Composable
@@ -85,20 +82,22 @@ private fun TransactionDetailsScreenContent(
             )
         },
         snackBar = { SnackBarContainer(snackBarState = state.snackBar) },
+        errorState = state.errorState,
+        onRetry = { interactionListener.onRefresh() }
     ) {
         Crossfade(
             targetState = state,
             modifier = Modifier.fillMaxSize()
         ) {
             when {
-                (state.isError != null) -> {
-                    NoInternetScreen(onRetry = interactionListener::onRefresh)
-                }
                 state.isLoading -> {
-                    Box(modifier = Modifier.fillMaxSize()){
+                    Box(modifier = Modifier.fillMaxSize()) {
                         ThreeDotsLoadingIndicator(modifier = Modifier.align(Alignment.Center))
                     }
                 }
+
+                state.errorState != null -> ErrorView(onRetry = { interactionListener.onRefresh() })
+
                 else -> {
                     Box {
                         val captureController = rememberCaptureController()
@@ -130,9 +129,7 @@ private fun onTransactionDetailsEffect(
     onNavigateBackClicked: () -> Unit,
 ) {
     when (effect) {
-        TransactionDetailsEffect.NavigateBack -> {
-            onNavigateBackClicked()
-        }
+        TransactionDetailsEffect.NavigateBack -> onNavigateBackClicked()
     }
 }
