+++ wallet_presentation/src/commonMain/kotlin/net/thechance/mena/wallet/presentation/screen/transaction_details/TransactionDetailsViewModel.kt
@@ -10,7 +10,8 @@ import mena.wallet_presentation.generated.resources.share_transaction_details_er
 import net.thechance.mena.wallet.domain.entity.Transaction
 import net.thechance.mena.wallet.domain.repository.TransactionRepository
 import net.thechance.mena.wallet.presentation.base.BaseViewModel
-import net.thechance.mena.wallet.presentation.base.SnackBarState
+import net.thechance.mena.wallet.presentation.base.ErrorState
+import net.thechance.mena.wallet.presentation.model.SnackBarState
 import net.thechance.mena.wallet.presentation.utils.ImageSharer
 import org.jetbrains.compose.resources.StringResource
 import org.koin.android.annotation.KoinViewModel
@@ -25,9 +26,8 @@ class TransactionDetailsViewModel(
     @Provided val transactionRepository: TransactionRepository,
     private val ioDispatcher: CoroutineDispatcher = Dispatchers.IO,
     @Provided private val transactionId: String,
-) : BaseViewModel<TransactionDetailsScreenState, TransactionDetailsEffect>(
-    TransactionDetailsScreenState()
-), TransactionDetailsInteractionListener {
+) : BaseViewModel<TransactionDetailsScreenState, TransactionDetailsEffect>
+    (TransactionDetailsScreenState()), TransactionDetailsInteractionListener {
 
     init {
         getTransactionDetails()
@@ -49,15 +49,12 @@ class TransactionDetailsViewModel(
 
     private fun onGetTransactionDetailsSuccess(transaction: Transaction) {
         updateState {
-            it.copy(
-                isLoading = false,
-                transactionDetailsUiState =transaction.toUi()
-            )
+            it.copy(isLoading = false, transactionDetailsUiState = transaction.toUi())
         }
     }
 
-    private fun onGetTransactionDetailsError(throwable: Throwable) {
-        updateState { it.copy(isLoading = false, isError = throwable) }
+    private fun onGetTransactionDetailsError(errorState: ErrorState) {
+        updateState { it.copy(isLoading = false, errorState = errorState) }
     }
 
     private fun onGetTransactionDetailsStart() {
@@ -103,7 +100,7 @@ class TransactionDetailsViewModel(
         updateState { it.copy(isShareReceiptBtnLoading = false) }
     }
 
-    private suspend fun onShareReceiptError(throwable: Throwable) {
+    private suspend fun onShareReceiptError(errorState: ErrorState) {
         updateState { it.copy(isShareReceiptBtnLoading = false) }
         showSnackBar(
             titleRes = Res.string.error,
@@ -147,7 +144,7 @@ class TransactionDetailsViewModel(
     }
 
     override fun onRefresh() {
-        updateState { it.copy(isLoading = true, isError = null) }
+        updateState { it.copy(isLoading = true, errorState = null) }
         getTransactionDetails()
     }
